refactor: Update setup-machine.sh
Try out a method of not defining an explicit version number for the bat package. Instead, get the latest version from the repository. In addition, check to see if it is already installed and/or with a newer version and interact with the user accordingly.

This logic will be carried forward to similar functions in subsequent commits.